% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K9_metrics.R
\name{k9_get_metrics}
\alias{k9_get_metrics}
\title{Query Time Series Points}
\usage{
k9_get_metrics(from = NULL, to = NULL, query)
}
\arguments{
\item{from}{seconds since the unix epoch}

\item{to}{seconds since the unix epoch}

\item{query}{query string}
}
\description{
This end point allows you to query for metrics from any time period.
}
\seealso{
\url{http://docs.datadoghq.com/graphing/}
}

