% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K9_metrics.R
\name{k9_list_active_metrics}
\alias{k9_list_active_metrics}
\title{Get List Of Active Metrics}
\usage{
k9_list_active_metrics(from = NULL)
}
\arguments{
\item{from}{seconds since the unix epoch}
}
\description{
Get the list of actively reporting metrics from a given time until now.
}
