% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K9_metrics.R
\name{k9_list_metrics}
\alias{k9_list_metrics}
\title{Get List Of Active Metrics}
\usage{
k9_list_metrics(from = NULL)
}
\arguments{
\item{from}{seconds since the unix epoch}
}
\description{
Get the list of actively reporting metrics from a given time until now.
}
\examples{
\dontrun{
# by default, list all metrics available on the current datetime
k9_list_metrics()

# if from argument is provided, this tries to get active metrics from the time
k9_list_metrics(Sys.Date() - 1)
}

}
