% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K9_metrics.R
\name{k9_post_metric}
\alias{k9_post_metric}
\title{Posts a metric value to Datadog}
\usage{
k9_post_metric(metric, metric_type, value, tags = list(), interval = NULL)
}
\arguments{
\item{metric}{the name of the time series}

\item{metric_type}{type of your metric either: gauge, rate, or count. Optional, default=gauge}

\item{value}{the numeric value to post}

\item{tags}{a list of tags associated with the metric.}

\item{interval}{if the type of the metric is rate or count, define the corresponding interval. Optional, default=None}
}
\description{
This end point allows you to post time-series data that can be graphed on Datadog’s dashboards or queried from any time period.
}
\details{
The Datadog API uses resource-oriented URLs, uses status codes to indicate the success or failure of requests and returns JSON from all requests.
With this method you can post counters, gauges to measure the value of a
particular thing over time and rates that represent the derivative of a
metric, it’s the value variation of a metric on a defined time interval.
}
\seealso{
\url{http://docs.datadoghq.com/api/?lang=console#metrics}
\url{http://docs.datadoghq.com/graphing/}
\url{https://docs.datadoghq.com/developers/metrics/#metric-types}
}
