% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K9_events.R
\name{k9_events}
\alias{k9_events}
\title{Get Events}
\usage{
k9_events(event_id = NULL, start = NULL, end = NULL, priority = c("low",
  "normal"), sources = NULL, tags = NULL)
}
\arguments{
\item{event_id}{Event ID}

\item{start}{POSIX timestamp}

\item{end}{POSIX timestamp}

\item{priority}{\code{low} or \code{normal}}

\item{sources}{A comma separated string of sources}

\item{tags}{A comma separated string of tags}
}
\description{
This end point allows you to query for event details.
}
\seealso{
\link{http://docs.datadoghq.com/api/?lang=console#events}
}
