% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K9_metrics.R
\name{k9_get_metrics}
\alias{k9_get_metrics}
\title{Query Time Series Points}
\usage{
k9_get_metrics(query = NULL, metric = NULL, scope = list(), by = NULL,
  from = NULL, to = NULL)
}
\arguments{
\item{query}{query string}

\item{metric}{metric name}

\item{scope}{list of scopes (\code{scope})}

\item{by}{key to group aggregation}

\item{from}{seconds since the unix epoch}

\item{to}{seconds since the unix epoch}
}
\description{
This end point allows you to query for metrics from any time period.
}
\seealso{
\url{http://docs.datadoghq.com/api/?lang=console#metrics}, \url{http://docs.datadoghq.com/graphing/}
}
